feat: QPPA-8977 add misc tests
