+++ scripts/measures/build-2018-measures
@@ -57,4 +57,4 @@ cat measures/$currentPerformanceYear/measures-data.json | \
 # 3. Validate that measures actually existed in their listed firstPerformanceYear and the years after
 echo "validating that measures actually existed in their listed firstPerformanceYear and the years after"
 cat measures/$currentPerformanceYear/measures-data.json | \
-	node scripts/validate-measures-past-existence.js $currentPerformanceYear
+	node scripts/measures/archive/validate-measures-past-existence.js $currentPerformanceYear
