+++ test/mvps/mvp-raw-test-bad-id.json
@@ -1,0 +1,29 @@
+[
+  {
+    "MVP Title": "Test Title 5",
+    "MVP ID": "M1368",
+    "MVP Description": "Test Description 5.",
+    "Most Applicable Medical Specialties": "Infectious Disease; Immunology",
+    "Clinical Topic": "Infectious Disease, Immunology",
+    "MVP Reporting Category": "Foundational",
+    "Measure Id": "PI_INFBLO_1"
+  },
+  {
+    "MVP Title": "Test Title 5",
+    "MVP ID": "M1368",
+    "MVP Description": "Test Description 5.",
+    "Most Applicable Medical Specialties": "Infectious Disease; Immunology",
+    "Clinical Topic": "Infectious Disease, Immunology",
+    "MVP Reporting Category": "Foundational",
+    "Measure Id": "PI_PHCDRR_2_PROD"
+  },
+  {
+    "MVP Title": "Test Title 5",
+    "MVP ID": "M1368",
+    "MVP Description": "Test Description 5.",
+    "Most Applicable Medical Specialties": "Infectious Disease; Immunology",
+    "Clinical Topic": "Infectious Disease, Immunology",
+    "MVP Reporting Category": "Foundational",
+    "Measure Id": "fakeId"
+  }
+]
