+++ test/mvps/mvp-raw-test.json
@@ -1,0 +1,164 @@
+[
+  {
+    "MVP Title": "Test Title 1",
+    "MVP ID": "G0053",
+    "MVP Description": "Test Description 1.",
+    "Most Applicable Medical Specialties": "Rheumatology",
+    "Clinical Topic": "Rheumatology",
+    "MVP Reporting Category": "Cost",
+    "Measure Id": "TPCC_1"
+  },
+  {
+    "MVP Title": "Test Title 1",
+    "MVP ID": "G0053",
+    "MVP Description": "Test Description 1.",
+    "Most Applicable Medical Specialties": "Rheumatology",
+    "Clinical Topic": "Rheumatology",
+    "MVP Reporting Category": "Foundational",
+    "Measure Id": "PI_INFBLO_1"
+  },
+  {
+    "MVP Title": "Test Title 1",
+    "MVP ID": "G0053",
+    "MVP Description": "Test Description 1.",
+    "Most Applicable Medical Specialties": "Rheumatology",
+    "Clinical Topic": "Rheumatology",
+    "MVP Reporting Category": "Foundational",
+    "Measure Id": "PI_PHCDRR_2_PROD"
+  },
+  {
+    "MVP Title": "Test Title 1",
+    "MVP ID": "G0053",
+    "MVP Description": "Test Description 1.",
+    "Most Applicable Medical Specialties": "Rheumatology",
+    "Clinical Topic": "Rheumatology",
+    "MVP Reporting Category": "Quality",
+    "Measure Id": "479"
+  },
+  {
+    "MVP Title": "Test Title 2",
+    "MVP ID": "G0054",
+    "MVP Description": "Test Description 2.",
+    "Most Applicable Medical Specialties": "Neurology; Neurosurgical; Vascular Surgery",
+    "Clinical Topic": "Stroke Care and Prevention",
+    "MVP Reporting Category": "Foundational",
+    "Measure Id": "PI_INFBLO_1"
+  },
+  {
+    "MVP Title": "Test Title 2",
+    "MVP ID": "G0054",
+    "MVP Description": "Test Description 2.",
+    "Most Applicable Medical Specialties": "Neurology; Neurosurgical; Vascular Surgery",
+    "Clinical Topic": "Stroke Care and Prevention",
+    "MVP Reporting Category": "Foundational",
+    "Measure Id": "PI_PHCDRR_2_PROD"
+  },
+  {
+    "MVP Title": "Test Title 2",
+    "MVP ID": "G0054",
+    "MVP Description": "Test Description 2.",
+    "Most Applicable Medical Specialties": "Neurology; Neurosurgical; Vascular Surgery",
+    "Clinical Topic": "Stroke Care and Prevention",
+    "MVP Reporting Category": "Quality",
+    "Measure Id": "479"
+  },
+  {
+    "MVP Title": "Test Title 3",
+    "MVP ID": "G0058",
+    "MVP Description": "Test Description 3.",
+    "Most Applicable Medical Specialties": "Orthopedic Surgery",
+    "Clinical Topic": "Lower Extremity Joint Repair",
+    "MVP Reporting Category": "Foundational",
+    "Measure Id": "470"
+  },
+  {
+    "MVP Title": "Test Title 3",
+    "MVP ID": "G0058",
+    "MVP Description": "Test Description 3.",
+    "Most Applicable Medical Specialties": "Orthopedic Surgery",
+    "Clinical Topic": "Lower Extremity Joint Repair",
+    "MVP Reporting Category": "Foundational",
+    "Measure Id": "PI_INFBLO_1"
+  },
+  {
+    "MVP Title": "Test Title 3",
+    "MVP ID": "G0058",
+    "MVP Description": "Test Description 3.",
+    "Most Applicable Medical Specialties": "Orthopedic Surgery",
+    "Clinical Topic": "Lower Extremity Joint Repair",
+    "MVP Reporting Category": "Foundational",
+    "Measure Id": "PI_PHCDRR_2_PROD"
+  },
+  {
+    "MVP Title": "Test Title 4",
+    "MVP ID": "M0001",
+    "MVP Description": "Test Description 4.",
+    "Most Applicable Medical Specialties": "Oncology; Hematology",
+    "Clinical Topic": "Oncology and Hematology",
+    "MVP Reporting Category": "Cost",
+    "Measure Id": "TPCC_1"
+  },
+  {
+    "MVP Title": "Test Title 4",
+    "MVP ID": "M0001",
+    "MVP Description": "Test Description 4.",
+    "Most Applicable Medical Specialties": "Oncology; Hematology",
+    "Clinical Topic": "Oncology and Hematology",
+    "MVP Reporting Category": "Foundational",
+    "Measure Id": "PI_INFBLO_1"
+  },
+  {
+    "MVP Title": "Test Title 4",
+    "MVP ID": "M0001",
+    "MVP Description": "Test Description 4.",
+    "Most Applicable Medical Specialties": "Oncology; Hematology",
+    "Clinical Topic": "Oncology and Hematology",
+    "MVP Reporting Category": "Foundational",
+    "Measure Id": "PI_PHCDRR_2_PROD"
+  },
+  {
+    "MVP Title": "Test Title 4",
+    "MVP ID": "M0001",
+    "MVP Description": "Test Description 4.",
+    "Most Applicable Medical Specialties": "Oncology; Hematology",
+    "Clinical Topic": "Oncology and Hematology",
+    "MVP Reporting Category": "Quality",
+    "Measure Id": "321"
+  },
+  {
+    "MVP Title": "Test Title 5",
+    "MVP ID": "M1368",
+    "MVP Description": "Test Description 5.",
+    "Most Applicable Medical Specialties": "Infectious Disease; Immunology",
+    "Clinical Topic": "Infectious Disease, Immunology",
+    "MVP Reporting Category": "Improvement",
+    "Measure Id": "IA_AHE_5"
+  },
+  {
+    "MVP Title": "Test Title 5",
+    "MVP ID": "M1368",
+    "MVP Description": "Test Description 5.",
+    "Most Applicable Medical Specialties": "Infectious Disease; Immunology",
+    "Clinical Topic": "Infectious Disease, Immunology",
+    "MVP Reporting Category": "Cost",
+    "Measure Id": "TPCC_1"
+  },
+  {
+    "MVP Title": "Test Title 5",
+    "MVP ID": "M1368",
+    "MVP Description": "Test Description 5.",
+    "Most Applicable Medical Specialties": "Infectious Disease; Immunology",
+    "Clinical Topic": "Infectious Disease, Immunology",
+    "MVP Reporting Category": "Foundational",
+    "Measure Id": "PI_INFBLO_1"
+  },
+  {
+    "MVP Title": "Test Title 5",
+    "MVP ID": "M1368",
+    "MVP Description": "Test Description 5.",
+    "Most Applicable Medical Specialties": "Infectious Disease; Immunology",
+    "Clinical Topic": "Infectious Disease, Immunology",
+    "MVP Reporting Category": "Foundational",
+    "Measure Id": "PI_PHCDRR_2_PROD"
+  }
+]
