+++ test/mvps/mvp-test.json
@@ -1,0 +1,127 @@
+[
+  {
+    "mvpId": "G0053",
+    "clinicalTopic": "Rheumatology",
+    "title": "Test Title 1",
+    "description": "Test Description 1.",
+    "specialtiesMostApplicableTo": [
+      "Rheumatology"
+    ],
+    "clinicalTopics": "Rheumatology",
+    "qualityMeasureIds": [],
+    "iaMeasureIds": [],
+    "costMeasureIds": [
+      "TPCC_1"
+    ],
+    "foundationPiMeasureIds": [
+      "PI_INFBLO_1",
+      "PI_PHCDRR_2_PROD"
+    ],
+    "foundationQualityMeasureIds": [],
+    "administrativeClaimsMeasureIds": [
+      "479"
+    ],
+    "hasCahps": false,
+    "hasOutcomeAdminClaims": true
+  },
+  {
+    "mvpId": "G0054",
+    "clinicalTopic": "Stroke Care and Prevention",
+    "title": "Test Title 2",
+    "description": "Test Description 2.",
+    "specialtiesMostApplicableTo": [
+      "Neurology",
+      "Neurosurgical",
+      "Vascular Surgery"
+    ],
+    "clinicalTopics": "Stroke Care and Prevention",
+    "qualityMeasureIds": [],
+    "iaMeasureIds": [],
+    "costMeasureIds": [],
+    "foundationPiMeasureIds": [
+      "PI_INFBLO_1",
+      "PI_PHCDRR_2_PROD"
+    ],
+    "foundationQualityMeasureIds": [],
+    "administrativeClaimsMeasureIds": [
+      "479"
+    ],
+    "hasCahps": false,
+    "hasOutcomeAdminClaims": true
+  },
+  {
+    "mvpId": "G0058",
+    "clinicalTopic": "Lower Extremity Joint Repair",
+    "title": "Test Title 3",
+    "description": "Test Description 3.",
+    "specialtiesMostApplicableTo": [
+      "Orthopedic Surgery"
+    ],
+    "clinicalTopics": "Lower Extremity Joint Repair",
+    "qualityMeasureIds": [],
+    "iaMeasureIds": [],
+    "costMeasureIds": [],
+    "foundationPiMeasureIds": [
+      "PI_INFBLO_1",
+      "PI_PHCDRR_2_PROD"
+    ],
+    "foundationQualityMeasureIds": [
+      "470"
+    ],
+    "administrativeClaimsMeasureIds": [],
+    "hasCahps": false,
+    "hasOutcomeAdminClaims": false
+  },
+  {
+    "mvpId": "M0001",
+    "clinicalTopic": "Oncology and Hematology",
+    "title": "Test Title 4",
+    "description": "Test Description 4.",
+    "specialtiesMostApplicableTo": [
+      "Oncology",
+      "Hematology"
+    ],
+    "clinicalTopics": "Oncology and Hematology",
+    "qualityMeasureIds": [
+      "321"
+    ],
+    "iaMeasureIds": [],
+    "costMeasureIds": [
+      "TPCC_1"
+    ],
+    "foundationPiMeasureIds": [
+      "PI_INFBLO_1",
+      "PI_PHCDRR_2_PROD"
+    ],
+    "foundationQualityMeasureIds": [],
+    "administrativeClaimsMeasureIds": [],
+    "hasCahps": true,
+    "hasOutcomeAdminClaims": false
+  },
+  {
+    "mvpId": "M1368",
+    "clinicalTopic": "Infectious Disease, Immunology",
+    "title": "Test Title 5",
+    "description": "Test Description 5.",
+    "specialtiesMostApplicableTo": [
+      "Infectious Disease",
+      "Immunology"
+    ],
+    "clinicalTopics": "Infectious Disease, Immunology",
+    "qualityMeasureIds": [],
+    "iaMeasureIds": [
+      "IA_AHE_5"
+    ],
+    "costMeasureIds": [
+      "TPCC_1"
+    ],
+    "foundationPiMeasureIds": [
+      "PI_INFBLO_1",
+      "PI_PHCDRR_2_PROD"
+    ],
+    "foundationQualityMeasureIds": [],
+    "administrativeClaimsMeasureIds": [],
+    "hasCahps": false,
+    "hasOutcomeAdminClaims": false
+  }
+]
